REFACTORED: unnecessary interfaces removed
