+++ src/main/java/de/uhd/ifi/se/decision/management/jira/rest/ConfigRest.java
@@ -699,4 +699,18 @@ public class ConfigRest {
 	/* Configuration for Consistency		*/
 	/*										*/
 	/* **************************************/
+	@Path("/setConsistencyActivated")
+	@POST
+	public Response setConsistencyActivated(@Context HttpServletRequest request,
+											@QueryParam("projectKey") String projectKey,
+											@QueryParam("isConsistencyActivated") String isActivatedString) {
+		Response response = this.checkRequest(request, projectKey, isActivatedString);
+		if (response != null) {
+			return response;
+		}
+		boolean isActivated = Boolean.valueOf(isActivatedString);
+		ConfigPersistenceManager.setConsistencyActivated(projectKey, isActivated);
+		return Response.ok(Status.ACCEPTED).build();
+	}
+
 }
